backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow>=10.4.0
torch==2.8.0
torchvision==0.23.0
timm==0.9.12
numpy==1.26.4
python-jose[cryptography]==3.3.0
